more renames and zapt updates. Fix build rule dependency
