+++ examples/chip-tool/commands/clusters/Commands.h
@@ -22,10 +22,10 @@
 #include <cstdint>
 
 #include "ModelCommand.h"
-#include "gen/CHIPClientCallbacks.h"
-#include "gen/CHIPClusters.h"
 #include <lib/core/CHIPSafeCasts.h>
 #include <support/BytesToHex.h>
+#include <zap-generated/CHIPClientCallbacks.h>
+#include <zap-generated/CHIPClusters.h>
 
 static void OnDefaultSuccessResponse(void * context)
 {
