+++ examples/chip-tool/templates/commands.zapt
@@ -5,8 +5,8 @@
 #include <cstdint>
 
 #include "ModelCommand.h"
-#include "gen/CHIPClientCallbacks.h"
-#include "gen/CHIPClusters.h"
+#include <zap-generated/CHIPClientCallbacks.h>
+#include <zap-generated/CHIPClusters.h>
 #include <lib/core/CHIPSafeCasts.h>
 #include <support/BytesToHex.h>
