+++ examples/lock-app/lock-common/BUILD.gn
@@ -29,6 +29,6 @@ if (chip_enable_pw_rpc) {
 chip_data_model("lock-common") {
   zap_file = "lock-app.zap"
 
-  zap_pregenerated_dir = "gen"
+  zap_pregenerated_dir = "zap-generated"
   is_server = true
 }
