+++ examples/pump-controller-app/pump-controller-common/BUILD.gn
@@ -19,6 +19,6 @@ import("${chip_root}/src/app/chip_data_model.gni")
 chip_data_model("pump-controller-common") {
   zap_file = "pump-controller-app.zap"
 
-  zap_pregenerated_dir = "gen"
+  zap_pregenerated_dir = "zap-generated"
   is_server = true
 }
