+++ src/controller/CHIPDevice.cpp
@@ -26,7 +26,7 @@
 
 #include <controller/CHIPDevice.h>
 
-#include <controller/data_model/gen/CHIPClusters.h>
+#include <controller/data_model/zap-generated/CHIPClusters.h>
 
 #if CONFIG_DEVICE_LAYER
 #include <platform/CHIPDeviceLayer.h>
