+++ src/controller/CHIPDeviceController.cpp
@@ -36,7 +36,7 @@
 #include <controller/CHIPDeviceController.h>
 
 #include <app/common/gen/enums.h>
-#include <controller/data_model/gen/CHIPClusters.h>
+#include <controller/data_model/zap-generated/CHIPClusters.h>
 
 #if CONFIG_DEVICE_LAYER
 #include <platform/CHIPDeviceLayer.h>
