+++ src/controller/data_model/BUILD.gn
@@ -19,8 +19,10 @@ import("${chip_root}/src/app/chip_data_model.gni")
 chip_data_model("data_model") {
   zap_file = "controller-clusters.zap"
 
-  zap_pregenerated_dir = "gen"
+  zap_pregenerated_dir = "zap-generated"
 
   use_default_client_callbacks = true
   allow_circular_includes_from = [ "${chip_root}/src/controller" ]
+
+  internal_controller_model = true
 }
