+++ src/controller/data_model/zap-generated/CHIPClientCallbacks.cpp
@@ -17,7 +17,7 @@
 
 // THIS FILE IS GENERATED BY ZAP
 
-#include "gen/CHIPClientCallbacks.h"
+#include <zap-generated/CHIPClientCallbacks.h>
 
 #include <cinttypes>
