+++ src/controller/java/gen/CHIPClusters-JNI.cpp
@@ -17,8 +17,8 @@
 
 // THIS FILE IS GENERATED BY ZAP
 
-#include "gen/CHIPClientCallbacks.h"
-#include "gen/CHIPClusters.h"
+#include <zap-generated/CHIPClientCallbacks.h>
+#include <zap-generated/CHIPClusters.h>
 
 #include <controller/java/CHIPJNIError.h>
 #include <controller/java/JniReferences.h>
